moving basic util functions to __init__
